pyproject.toml
[project]
name = "file-parse-engine"
version = "0.2.0"
description = "VLM-powered multi-format file parsing engine for RAG pipelines"
readme = "README.md"
authors = [
    { name = "linzhiqin", email = "13818524101@163.com" }
]
requires-python = ">=3.11"
dependencies = [
    # Core
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    # CLI
    "typer>=0.12",
    "rich>=13.0",
    # Parsers
    "pymupdf>=1.24",
    "python-docx>=1.0",
    "python-pptx>=1.0",
    "openpyxl>=3.1",
    "pillow>=10.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "readability-lxml>=0.8",
    "ebooklib>=0.18",
    "striprtf>=0.0.26",
]

[project.scripts]
fpe = "file_parse_engine.cli:app"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[project.optional-dependencies]
ocr = [
    "rapidocr>=3.7",
    "onnxruntime>=1.19",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
